Rough setup 3. Added midi deserializing and started svg generating

+++ src/music/music_box/mod.rs
@@ -1,20 +1,21 @@
 // Modules
 pub mod functions;
 
-use serde_derive::{Deserialize, Serialize};
+// serde_derive
+use serde_derive::Deserialize;
 
 // Internal
 use super::note::Note;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Deserialize)]
 pub struct MusicBox {
     pub name: String,
-    pub strip_height_mm: u32,
+    pub strip_height_mm: f32,
     notes: Vec<Note>,
 }
 
 impl MusicBox {
-    pub fn new(name: String, strip_height_mm: u32, notes: Vec<Note>) -> Self {
+    pub fn new(name: String, strip_height_mm: f32, notes: Vec<Note>) -> Self {
         Self {
             name,
             strip_height_mm,
